refactor: create seperate types files for store-slice
