+++ client/app/components/types/store/slice/chat.ts
@@ -1,0 +1,7 @@
+import { Message as MessageInterface } from "../../chat";
+
+export type ChatSlice = {
+  projectChat: MessageInterface[];
+  addMessageInProjectChat: (message: MessageInterface) => void;
+  clearChat: () => void;
+};
