+++ client/app/components/types/store/slice/socket.ts
@@ -1,0 +1,8 @@
+import { Socket } from "socket.io-client";
+
+export type SocketSlice = {
+  socket: Socket | null;
+  connectSocket: () => void;
+  isConnectedToServer: boolean;
+  updateIsConnectedToServer: (value: boolean) => void;
+};
