+++ client/app/components/types/store/slice/user.ts
@@ -1,0 +1,9 @@
+export type UserSlice = {
+    showWelcomeScreen: boolean;
+    setShowWelcomeScreen: (value: boolean) => void;
+    currentUsername: string;
+    updatedCurrentUsername: (username: string) => void;
+    currentProjectName: string;
+    updateCurrentProjectName: (projectName: string) => void;
+  };
+
