+++ client/components/store/slices/chat-slice.ts
@@ -1,4 +1,4 @@
-import { ChatSlice, SetStateType } from "@/app/components/types/store";
+import { ChatSlice, SetStateType } from "@/app/components/types/store/store";
 
 export const createChatSlice = (set: SetStateType): ChatSlice => ({
   projectChat: [],
