+++ client/components/store/slices/project-slice.ts
@@ -1,4 +1,4 @@
-import { SetStateType, ProjectSlice } from "@/app/components/types/store";
+import { SetStateType, ProjectSlice } from "@/app/components/types/store/store";
 import { DEFAULT_PROJECT_STRUCTURE } from "@/lib/constants";
 
 export const createProjectSlice = (set: SetStateType): ProjectSlice => ({
@@ -49,5 +49,30 @@ export const createProjectSlice = (set: SetStateType): ProjectSlice => ({
   updateProjectClientsList: (updatedList) =>
     set(() => ({
       projectClientsList: updatedList,
+    })),
+
+  userCreatedProjectsList: [],
+  userJoinedProjectsList: [],
+
+  addProjectinCreatedProjectsList: (newProject) =>
+    set((state) => ({
+      userCreatedProjectsList: [newProject, ...state.userCreatedProjectsList],
+    })),
+  addProjectinJoinedProjectsList: (newProject) =>
+    set((state) => ({
+      userJoinedProjectsList: [newProject, ...state.userJoinedProjectsList],
+    })),
+
+  removeProjectinCreatedProjectsList: (projectId) =>
+    set((state) => ({
+      userCreatedProjectsList: state.userCreatedProjectsList.filter(
+        (projects) => projects.projectId !== projectId,
+      ),
+    })),
+  removeProjectinJoinedProjectsList: (projectId) =>
+    set((state) => ({
+      userJoinedProjectsList: state.userJoinedProjectsList.filter(
+        (projects) => projects.projectId !== projectId,
+      ),
     })),
 });
