+++ client/components/store/slices/socket-slice.ts
@@ -1,4 +1,4 @@
-import { SetStateType, SocketSlice } from "@/app/components/types/store";
+import { SetStateType, SocketSlice } from "@/app/components/types/store/store";
 import { io } from "socket.io-client";
 
 const SOCKET_SERVER_URL = process.env.NEXT_PUBLIC_SOCKET_URL!;
