+++ client/components/store/slices/user-slice.ts
@@ -1,4 +1,4 @@
-import { SetStateType, UserSlice } from "@/app/components/types/store";
+import { SetStateType, UserSlice } from "@/app/components/types/store/store";
 
 export const createUserSlice = (set: SetStateType): UserSlice => ({
   showWelcomeScreen: true,
@@ -20,28 +20,5 @@ export const createUserSlice = (set: SetStateType): UserSlice => ({
       currentProjectName: projectName,
     })),
 
-  userCreatedProjectsList: [],
-  userJoinedProjectsList: [],
-
-  addProjectinCreatedProjectsList: (newProject) =>
-    set((state) => ({
-      userCreatedProjectsList: [newProject, ...state.userCreatedProjectsList],
-    })),
-  addProjectinJoinedProjectsList: (newProject) =>
-    set((state) => ({
-      userJoinedProjectsList: [newProject, ...state.userJoinedProjectsList],
-    })),
-
-  removeProjectinCreatedProjectsList: (projectId) =>
-    set((state) => ({
-      userCreatedProjectsList: state.userCreatedProjectsList.filter(
-        (projects) => projects.projectId !== projectId,
-      ),
-    })),
-  removeProjectinJoinedProjectsList: (projectId) =>
-    set((state) => ({
-      userJoinedProjectsList: state.userJoinedProjectsList.filter(
-        (projects) => projects.projectId !== projectId,
-      ),
-    })),
+  
 });
